(recordTime) : Print the elapsed time since the launch
git-svn-id: http://es-operating-system.googlecode.com/svn/trunk/escort@2945 d15c3b8f-5b43-0410-be93-4d5a55ce2954

+++ src/Test.util.h
@@ -41,8 +41,7 @@ void eval(org::w3c::dom::Node node);
 org::w3c::dom::Document loadDocument(std::istream& stream);
 org::w3c::dom::Document loadDocument(const char* html);
 
-double recordTime(const char* msg, ...);
-
+unsigned recordTime(const char* msg, ...);
 unsigned getTick();
 
 inline bool isReplacedElement(org::w3c::dom::Element& element)
